Games: Add scripted Ernest Hemingway
I have finally finished the scripted version of Ernest Hemingway, due to popular demand of the game. Due to the nature of having many entries and votes (and PMs) to have to worry about, I figured it'd be a great time to finally make use of the game system and have The Scribe worry about it all. This also updates the current codebase.